# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HackWerk HTTP API
  version: 0.4.0
  description: Interner Vertrag des serverseitig gerenderten HackWerk-Webdienstes.
servers:
  - url: http://localhost:18533
paths:
  /login:
    get:
      operationId: getLogin
      responses:
        "200": { description: Deutsches Anmeldeformular. }
    post:
      operationId: postLogin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password, writeOnly: true }
      responses:
        "303": { description: Anmeldung erfolgreich; opaque Session wurde rotiert. }
        "401": { description: Generische ungültige Zugangsdaten. }
        "429": { description: Loginlimit erreicht, mit derselben generischen sichtbaren Meldung. }
  /logout:
    post:
      operationId: postLogout
      security: [{ sessionCookie: [] }]
      responses:
        "303": { description: Session serverseitig widerrufen. }
        "403": { description: CSRF-Prüfung fehlgeschlagen. }
  /password:
    get:
      operationId: getPasswordChange
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Passwortänderungsformular. }
    post:
      operationId: postPasswordChange
      security: [{ sessionCookie: [] }]
      responses:
        "303": { description: Passwort geändert und bestehende Sessions widerrufen. }
        "403": { description: CSRF-Prüfung fehlgeschlagen. }
        "422": { description: Passwortpolicy oder Bestätigung ungültig. }
  /admin/users:
    get:
      operationId: listUsers
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Adminliste der Benutzer und Fahrerzuordnung. }
        "403": { description: Adminberechtigung fehlt. }
    post:
      operationId: createUser
      security: [{ sessionCookie: [] }]
      responses:
        "303": { description: Benutzer atomar angelegt und auditiert. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackplan_session